implementing functionality

+++ src/cmsc420/schema/spatial/Seedling.java
@@ -10,5 +10,5 @@ package cmsc420.schema.spatial;
  */
 public interface Seedling {
 
-	public SpatialStructure generate(float width, float height);
+	public SpatialStructure generate(int width, int height);
 }
